Added Enemy.cs
This is a functional version of the enemy with a finite state machine.
The states are:
patrolling, (Moves from node to node)
chasing, (Chases the last known position of it's vision target)
searching, (Currently just rotates, but this will be where a more involved behaviour will go)
returning, (Returns to the current node index in it's patrol path)
waiting (Waits at a node)

This currently does not include the varying types of patrol paths.

The scene includes an example of how the enemy will react to a player object.