Protoype Change Movement
Included models in the movement scene.

Small change to chasing.cs, where instead of a wait timer to decide when this will move, the chaser adds a function call to the arrive delegate of the target. This will count until the max number of nodes reached and then start chasing.