Added Layer Masks to vision
TargetMask = colliders with this mask are found as visibleTargets
obstacleMask = colliders with this mask will block raycasts.

Initially all colliders are found with an OverlapSphere using the vision radius and a targetMask.

Then a raycast is shot to each target found, if there is an obstacle that is found then they have not been seen, otherwise they are added to the visibleTargets list.

currently there are some calculations to find the tangent point to the player radius. The angle found from this vector is added to the vision cone to allow for the cone to register that the collider has crossed the vision threshold.

There is a simple addition to the enemy script. that if the vision.visible targets > 0; Then the player is spotted. There is currently only one player in the level, this is currently, ok, but it isn't the most optimal way to do this.