Multiple Raycasts are shot if target is semi behind an obstacle
Multiple raycasts are shot in decrementing angles from the centre target line to the tangent line of the targets sphere collider. This is an attempt to get as close to the edge as possible before deeming it as the target being not visible.

I'm pretty sure that this is not correctly implemented yet but it does happen to function due to the amount of raycasts that were shot.