Edge of Vision Checks
Many checks are made to see if the target is crossing the edge of the vision range and it is being partly obscured by obstacles.

The end result should be that the target is only seen when any part of it's collider is inside the vision radius without an obstacle in the way.