Code Cleanup and Gizmo draw: raycasts
Cleaned up code that was unnecessary.

Copied the code that finds the visible targets and added it to onGizmoDraw with drawing the raycasts.
This is a very inefficient way way to do this but it is simply a bandaid solution for the moment. I need to come back to this code later as vision needs to be drawn properly during runtime.