Collision With spherecast
A spherecast is made in the direction that the player is heading. If that cast hits an obstacle then a second sphere cast is sent in a perpendicular direction to the hit point and if that second cast doesn't hit an obstacle as well, then the player will move in that perpendicular direction.