Added Pause Mechanic
GameManager needed a new enum state to keep track of if the game is running or in a paused state.

Simply put this includes functions that will set the timeScale to zero when the game should be paused. All deltaTime related functions (eg. movement) will be multiplied by the timeScale, if it is zero then the result will also be zero.