Added PatrolNode.cs and prefab
This is a script that currently visually shows connections between nodes. But this may be able to be used later for more specifc node purposes, such as using a node as a trigger for example.

Each node stores a reference to an array of nodes that it is connected to. This will be able to be used as a direction to look for enemy search behaviours