Enemy: Exit chase find direction to search
When an enemy loses exits chase state it will find a direction to move while it searches. The initial (and currently only) direction is found by comparing closest Node connections to the last Known player heading that the enemy witnessed. The connection direction that is most similar to the last player heading is the direction that the enemy will follow. In the event that the most similar connection is the opposite direction to the last player heading, than the enemy will attempt to find a location simply in the players heading.