Enemy: Extra wait times at nodes
Enemies will make use of extra node wait times and look directions where an enemy will arrive at a node, wait for a specified time, then turn towards the specified look direction. Repeat until that patrol node does not have any more extra look directions.

When the enemy is ready to move on to the next patrol, it will enter a new state nodeExit. This is a simple timer that will wait a specified amount of time before moving on to the next node.