Enemy Search Behaviour makes use of point of interest nodes
The enemy will set their destination to a point of interest node while searching if it is in a specified radius. They will then set their destination to the first transform in the location array. When they arrive there they will continue to the next transform. And continue until the end.