Added CheckPoint as a Trigger Component
ReOrganised folder directory.

When The player enters this trigger. The GameManager's static Vector3 spawnPos is set to this checkPoint's spawnPosition. It will also change the GameManager's static bool checkPointActivated to true.

A new mthod was needed in the menuController to reload the current Scene. This function is simply called ReloadCurrentScene. If the hardReset bool is true, then The GameManager will not remember if the player has hit a checkpoint.