Added in a mouse control mode for camera
A ray is hot from the mouse cursor position in the main camera to the floor to find a point in world space. The vector from this point to the player is multiplied by a scalar to REDUCE the length and is then applied to the "cameraAxis" parent to move the location of the camera.