Fixed Player Collision
The player was not colliding with walls correctly. The notable problems were that wrong values were being used and applied. As a brief example, The second raycast from a wall was using a magnitude of a dotproduct instead of the speed he should be travelling. Hopefully all issues have been resolved.