Enemy correctly returns back to Patrol Nodes from Very Far Distance
The calculation for paths that were very far away were taking more than one frame before the path was finished. This meant that the remaining distance checks were not working correctly as it didn't quite know how far it was yet. A simple check of agent.pathPending will find if the path is still being calculated.